requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.26
django-ninja==1.5.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.1.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.46
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pluggy==1.6.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
xxhash==3.6.0
zstandard==0.25.0
